Added date arguments to plot and split tool
Currently date is used to set the start and end years of the output
files, but this could be changed to actually constrain the data to a
target date range.  The workspace is now set as a positional argument
but the script will assume that it is being called in the project folder
if a workspace is not set.