Using a debug flag to control debug level printing
The large number of logging.debug() statements were causing the code to
run quite slow.  It seems the strings were being built even if the
logging level was not debug.  Eventually build the loggers and set the
debug flag in a logging class or in the Data class.  I also set the
default format to float for most of the debug statements.