Floating point planting or greenup "date" causes error
The default date_for_pl_or_gu for canola was 4.833, which was causing an
error when computing the planting or greenup DOY.  I'm not sure why/how
the vb code is working.  I moved some of this code to the
crop_parameters.py since it only needs to be computed once per crop, not
for each time step.