Updates for GDD computation for winter crops
Instead of checking if a crop is a winter crop for each iteration, check
for "Winter" crops when initializing the crop and set a flag.  Tommy was
checking if "Winter" was in the crop name, but switched back to checking
for "Winter Wheat" explicitly to match VB code for now.