Bug fix - dcuttings and bcuttings variables
In the original code, it wasn't obvious if dcuttings and bcutting were
the beef and dairy cutting values from MeanCuttings.txt.  The code was
previously modified with the assumption they they are.  The logging
statement was still looking for the bcuttings and dcuttings variables.
Also some formatting changes to get the line lengths down under 80
characters.