Fixing debug logging issue
Before I was printing "crop" which was using the simple __str__ method
on the crop_params.  I updated the crop_params __str__ method to print
all the attributes but this means I need to print crop.name to just get
the crop name instead.

Also, I switched to using the string format method in crop_cycle.
Eventually I will change all of the code over to this.