Don't allow stress even to occur on first day of season closes #12
The issue came up in Carson County, TX 2011 for corn.  The stress
flag
was getting set on the first day of the season because it was windy
and
dry (leading up to that day and on that day), which caused Kc_bas to
go
just above 0.3 threshold.  The stress event was forcing Ks to 0,
which
caused Kc_act to be 0 for the rest of the season.  The stress
event was
also being set before the first irrigation was applied.  This
changes
makes it so a stress event can not occur on the first day of the
growing
season.