Adding additional flag to see if growing season ever goes active
The goal is to be able to easily identify and log any year where the
growing season never goes active.  This approach adds an additional flag
that is set True when the season starts and is cleared on the last day
of the year or 1 day before the start of winter crops (this is hardcoded
to Oct 1st/DOY 274 right now).  Since in_season is getting saved each
day, it might make more sense to just look "back" for more than 365 days
of in_season == 0.