Reclamation Revisions
Added ability to optionally use a configuration file.  This enables user
to specify parameters and units.  Original behavoir with hard-wired
parameters and units is still supported.  If value of '-i' argument ends
with '.ini' or 'cfg', it is used as a configuraiton file.  Otherwise, it
is used file data file name as per original usage.

Enabled use of time series data from Excel or csv files and provided
example files of both formats.

Updated documentation and creaed pdf version.