Updating documentation
Moved some of the text from the Example description to separate
sections.  I started a separate Data section that to try to explain what
the different ancillary datasets are for, a prep section to explain how
to run the prep tools and what the different options do, and then
simplified the CropET example section to only list the commands to run.
This is still a work in progress.